+++ Vocaluxe/Lib/Video/CVideoDecoderFFmpeg.cs
@@ -661,7 +661,7 @@ namespace Vocaluxe.Lib.Video
 
                 if (num >= 0)
                 {
-                    if (frame.Index == -1 || _Width != (int)frame.Width || _Height != (int)frame.Height)
+                    if (frame.Index == -1 || _Width != frame.Width || _Height != frame.Height)
                     {
                         CDraw.RemoveTexture(ref frame);
                         frame = CDraw.AddTexture(_Width, _Height, _FrameBuffer[num].Data);
